fix: define PostHog.init before Rails initializers
