+++ .changeset/early-rails-init.md
@@ -1,0 +1,5 @@
+---
+"posthog-ruby": patch
+---
+
+Fix Rails initializer load order so `PostHog.init` is available when `posthog-rails` is required.
